Type Definition packed_simd_2::f64x2

type f64x2 = Simd<[f64; 2]>;

A 128-bit vector with 2 f64 lanes.

Implementations

impl f64x2

pub const fn new(x0: f64, x1: f64) -> Self

Creates a new instance with each vector elements initialized with the provided values.

pub const fn lanes() -> usize

Returns the number of vector lanes.

pub const fn splat(value: f64) -> Self

Constructs a new instance with each element initialized to value.

pub fn extract(self, index: usize) -> f64

Extracts the value at index.

Panics

If index >= Self::lanes().

pub unsafe fn extract_unchecked(self, index: usize) -> f64

Extracts the value at index.

Safety

If index >= Self::lanes() the behavior is undefined.

#[must_use = "replace does not modify the original value - \ it returns a new vector with the value at `index` \ replaced by `new_value`d"]
pub fn replace(self, index: usize, new_value: f64) -> Self

Returns a new vector where the value at index is replaced by new_value.

Panics

If index >= Self::lanes().

#[must_use = "replace_unchecked does not modify the original value - \ it returns a new vector with the value at `index` \ replaced by `new_value`d"]
pub unsafe fn replace_unchecked(self, index: usize, new_value: f64) -> Self

Returns a new vector where the value at index is replaced by new_value.

Safety

If index >= Self::lanes() the behavior is undefined.

impl f64x2

pub fn min(self, x: Self) -> Self

Minimum of two vectors.

Returns a new vector containing the minimum value of each of the input vector lanes.

pub fn max(self, x: Self) -> Self

Maximum of two vectors.

Returns a new vector containing the maximum value of each of the input vector lanes.

impl f64x2

pub fn sum(self) -> f64

Horizontal sum of the vector elements.

The intrinsic performs a tree-reduction of the vector elements. That is, for an 8 element vector:

((x0 + x1) + (x2 + x3)) + ((x4 + x5) + (x6 + x7))

If one of the vector element is NaN the reduction returns NaN. The resulting NaN is not required to be equal to any of the NaNs in the vector.

pub fn product(self) -> f64

Horizontal product of the vector elements.

The intrinsic performs a tree-reduction of the vector elements. That is, for an 8 element vector:

((x0 * x1) * (x2 * x3)) * ((x4 * x5) * (x6 * x7))

If one of the vector element is NaN the reduction returns NaN. The resulting NaN is not required to be equal to any of the NaNs in the vector.

impl f64x2

pub fn max_element(self) -> f64

Largest vector element value.

pub fn min_element(self) -> f64

Smallest vector element value.

impl f64x2

pub fn from_slice_aligned(slice: &[f64]) -> Self

Instantiates a new vector with the values of the slice.

Panics

If slice.len() < Self::lanes() or &slice[0] is not aligned to an align_of::<Self>() boundary.

pub fn from_slice_unaligned(slice: &[f64]) -> Self

Instantiates a new vector with the values of the slice.

Panics

If slice.len() < Self::lanes().

pub unsafe fn from_slice_aligned_unchecked(slice: &[f64]) -> Self

Instantiates a new vector with the values of the slice.

Safety

If slice.len() < Self::lanes() or &slice[0] is not aligned to an align_of::<Self>() boundary, the behavior is undefined.

pub unsafe fn from_slice_unaligned_unchecked(slice: &[f64]) -> Self

Instantiates a new vector with the values of the slice.

Safety

If slice.len() < Self::lanes() the behavior is undefined.

impl f64x2

pub fn write_to_slice_aligned(self, slice: &mut [f64])

Writes the values of the vector to the slice.

Panics

If slice.len() < Self::lanes() or &slice[0] is not aligned to an align_of::<Self>() boundary.

pub fn write_to_slice_unaligned(self, slice: &mut [f64])

Writes the values of the vector to the slice.

Panics

If slice.len() < Self::lanes().

pub unsafe fn write_to_slice_aligned_unchecked(self, slice: &mut [f64])

Writes the values of the vector to the slice.

Safety

If slice.len() < Self::lanes() or &slice[0] is not aligned to an align_of::<Self>() boundary, the behavior is undefined.

pub unsafe fn write_to_slice_unaligned_unchecked(self, slice: &mut [f64])

Writes the values of the vector to the slice.

Safety

If slice.len() < Self::lanes() the behavior is undefined.

impl f64x2

pub fn shuffle1_dyn<I>(self, indices: I) -> Self where
Self: Shuffle1Dyn<Indices = I>,

Shuffle vector elements according to indices.

impl f64x2

pub const EPSILON: f64x2

Machine epsilon value.

pub const MIN: f64x2

Smallest finite value.

pub const MIN_POSITIVE: f64x2

Smallest positive normal value.

pub const MAX: f64x2

Largest finite value.

pub const NAN: f64x2

Not a Number (NaN).

pub const INFINITY: f64x2

Infinity (∞).

pub const NEG_INFINITY: f64x2

Negative infinity (-∞).

pub const PI: f64x2

Archimedes’ constant (π)

pub const FRAC_PI_2: f64x2

π/2

pub const FRAC_PI_3: f64x2

π/3

pub const FRAC_PI_4: f64x2

π/4

pub const FRAC_PI_6: f64x2

π/6

pub const FRAC_PI_8: f64x2

π/8

pub const FRAC_1_PI: f64x2

1/π

pub const FRAC_2_PI: f64x2

2/π

pub const FRAC_2_SQRT_PI: f64x2

2/sqrt(π)

pub const SQRT_2: f64x2

sqrt(2)

pub const FRAC_1_SQRT_2: f64x2

1/sqrt(2)

pub const E: f64x2

Euler’s number (e)

pub const LOG2_E: f64x2

log2(e)

pub const LOG10_E: f64x2

log10(e)

pub const LN_2: f64x2

ln(2)

pub const LN_10: f64x2

ln(10)

impl f64x2

pub fn is_nan(self) -> m64x2

pub fn is_infinite(self) -> m64x2

pub fn is_finite(self) -> m64x2

impl f64x2

pub fn abs(self) -> Self

Absolute value.

impl f64x2

pub fn cos(self) -> Self

Cosine.

pub fn cos_pi(self) -> Self

Cosine of self * PI.

impl f64x2

pub fn exp(self) -> Self

Returns the exponential function of self: e^(self).

impl f64x2

pub fn ln(self) -> Self

Returns the natural logarithm of self.

impl f64x2

pub fn mul_add(self, y: Self, z: Self) -> Self

Fused multiply add: self * y + z

impl f64x2

pub fn mul_adde(self, y: Self, z: Self) -> Self

Fused multiply add estimate: ~= self * y + z

While fused multiply-add (fma) has infinite precision, mul_adde has at worst the same precision of a multiply followed by an add. This might be more efficient on architectures that do not have an fma instruction.

impl f64x2

pub fn powf(self, x: Self) -> Self

Raises self number to the floating point power of x.

impl f64x2

pub fn recpre(self) -> Self

Reciprocal estimate: ~= 1. / self.

FIXME: The precision of the estimate is currently unspecified.

impl f64x2

pub fn rsqrte(self) -> Self

Reciprocal square-root estimate: ~= 1. / self.sqrt().

FIXME: The precision of the estimate is currently unspecified.

impl f64x2

pub fn sin(self) -> Self

Sine.

pub fn sin_pi(self) -> Self

Sine of self * PI.

pub fn sin_cos_pi(self) -> (Self, Self)

Sine and cosine of self * PI.

impl f64x2

pub fn sqrt(self) -> Self

impl f64x2

pub fn sqrte(self) -> Self

Square-root estimate.

FIXME: The precision of the estimate is currently unspecified.

impl f64x2

pub fn tanh(self) -> Self

Tanh.

impl f64x2

pub fn eq(self, other: Self) -> m64x2

Lane-wise equality comparison.

pub fn ne(self, other: Self) -> m64x2

Lane-wise inequality comparison.

pub fn lt(self, other: Self) -> m64x2

Lane-wise less-than comparison.

pub fn le(self, other: Self) -> m64x2

Lane-wise less-than-or-equals comparison.

pub fn gt(self, other: Self) -> m64x2

Lane-wise greater-than comparison.

pub fn ge(self, other: Self) -> m64x2

Lane-wise greater-than-or-equals comparison.