Type Definition packed_simd_2::i8x64

type i8x64 = Simd<[i8; 64]>;

A 512-bit vector with 64 i8 lanes.

Implementations

impl i8x64

pub const fn new(
x0: i8,
x1: i8,
x2: i8,
x3: i8,
x4: i8,
x5: i8,
x6: i8,
x7: i8,
x8: i8,
x9: i8,
x10: i8,
x11: i8,
x12: i8,
x13: i8,
x14: i8,
x15: i8,
x16: i8,
x17: i8,
x18: i8,
x19: i8,
x20: i8,
x21: i8,
x22: i8,
x23: i8,
x24: i8,
x25: i8,
x26: i8,
x27: i8,
x28: i8,
x29: i8,
x30: i8,
x31: i8,
x32: i8,
x33: i8,
x34: i8,
x35: i8,
x36: i8,
x37: i8,
x38: i8,
x39: i8,
x40: i8,
x41: i8,
x42: i8,
x43: i8,
x44: i8,
x45: i8,
x46: i8,
x47: i8,
x48: i8,
x49: i8,
x50: i8,
x51: i8,
x52: i8,
x53: i8,
x54: i8,
x55: i8,
x56: i8,
x57: i8,
x58: i8,
x59: i8,
x60: i8,
x61: i8,
x62: i8,
x63: i8
) -> Self

Creates a new instance with each vector elements initialized with the provided values.

pub const fn lanes() -> usize

Returns the number of vector lanes.

pub const fn splat(value: i8) -> Self

Constructs a new instance with each element initialized to value.

pub fn extract(self, index: usize) -> i8

Extracts the value at index.

Panics

If index >= Self::lanes().

pub unsafe fn extract_unchecked(self, index: usize) -> i8

Extracts the value at index.

Safety

If index >= Self::lanes() the behavior is undefined.

#[must_use = "replace does not modify the original value - \ it returns a new vector with the value at `index` \ replaced by `new_value`d"]
pub fn replace(self, index: usize, new_value: i8) -> Self

Returns a new vector where the value at index is replaced by new_value.

Panics

If index >= Self::lanes().

#[must_use = "replace_unchecked does not modify the original value - \ it returns a new vector with the value at `index` \ replaced by `new_value`d"]
pub unsafe fn replace_unchecked(self, index: usize, new_value: i8) -> Self

Returns a new vector where the value at index is replaced by new_value.

Safety

If index >= Self::lanes() the behavior is undefined.

impl i8x64

pub fn rotate_left(self, n: i8x64) -> i8x64

Shifts the bits of each lane to the left by the specified amount in the corresponding lane of n, wrapping the truncated bits to the end of the resulting integer.

Note: this is neither the same operation as << nor equivalent to slice::rotate_left.

pub fn rotate_right(self, n: i8x64) -> i8x64

Shifts the bits of each lane to the right by the specified amount in the corresponding lane of n, wrapping the truncated bits to the beginning of the resulting integer.

Note: this is neither the same operation as << nor equivalent to slice::rotate_left.

impl i8x64

pub fn min(self, x: Self) -> Self

Minimum of two vectors.

Returns a new vector containing the minimum value of each of the input vector lanes.

pub fn max(self, x: Self) -> Self

Maximum of two vectors.

Returns a new vector containing the maximum value of each of the input vector lanes.

impl i8x64

pub fn wrapping_sum(self) -> i8

Horizontal wrapping sum of the vector elements.

The intrinsic performs a tree-reduction of the vector elements. That is, for an 8 element vector:

((x0 + x1) + (x2 + x3)) + ((x4 + x5) + (x6 + x7))

If an operation overflows it returns the mathematical result modulo 2^n where n is the number of times it overflows.

pub fn wrapping_product(self) -> i8

Horizontal wrapping product of the vector elements.

The intrinsic performs a tree-reduction of the vector elements. That is, for an 8 element vector:

((x0 * x1) * (x2 * x3)) * ((x4 * x5) * (x6 * x7))

If an operation overflows it returns the mathematical result modulo 2^n where n is the number of times it overflows.

impl i8x64

pub fn max_element(self) -> i8

Largest vector element value.

pub fn min_element(self) -> i8

Smallest vector element value.

impl i8x64

pub fn and(self) -> i8

Lane-wise bitwise and of the vector elements.

Note: if the vector has one lane, the first element of the vector is returned.

pub fn or(self) -> i8

Lane-wise bitwise or of the vector elements.

Note: if the vector has one lane, the first element of the vector is returned.

pub fn xor(self) -> i8

Lane-wise bitwise xor of the vector elements.

Note: if the vector has one lane, the first element of the vector is returned.

impl i8x64

pub fn from_slice_aligned(slice: &[i8]) -> Self

Instantiates a new vector with the values of the slice.

Panics

If slice.len() < Self::lanes() or &slice[0] is not aligned to an align_of::<Self>() boundary.

pub fn from_slice_unaligned(slice: &[i8]) -> Self

Instantiates a new vector with the values of the slice.

Panics

If slice.len() < Self::lanes().

pub unsafe fn from_slice_aligned_unchecked(slice: &[i8]) -> Self

Instantiates a new vector with the values of the slice.

Safety

If slice.len() < Self::lanes() or &slice[0] is not aligned to an align_of::<Self>() boundary, the behavior is undefined.

pub unsafe fn from_slice_unaligned_unchecked(slice: &[i8]) -> Self

Instantiates a new vector with the values of the slice.

Safety

If slice.len() < Self::lanes() the behavior is undefined.

impl i8x64

pub fn write_to_slice_aligned(self, slice: &mut [i8])

Writes the values of the vector to the slice.

Panics

If slice.len() < Self::lanes() or &slice[0] is not aligned to an align_of::<Self>() boundary.

pub fn write_to_slice_unaligned(self, slice: &mut [i8])

Writes the values of the vector to the slice.

Panics

If slice.len() < Self::lanes().

pub unsafe fn write_to_slice_aligned_unchecked(self, slice: &mut [i8])

Writes the values of the vector to the slice.

Safety

If slice.len() < Self::lanes() or &slice[0] is not aligned to an align_of::<Self>() boundary, the behavior is undefined.

pub unsafe fn write_to_slice_unaligned_unchecked(self, slice: &mut [i8])

Writes the values of the vector to the slice.

Safety

If slice.len() < Self::lanes() the behavior is undefined.

impl i8x64

pub fn swap_bytes(self) -> Self

Reverses the byte order of the vector.

pub fn to_le(self) -> Self

Converts self to little endian from the target’s endianness.

On little endian this is a no-op. On big endian the bytes are swapped.

pub fn to_be(self) -> Self

Converts self to big endian from the target’s endianness.

On big endian this is a no-op. On little endian the bytes are swapped.

pub fn from_le(x: Self) -> Self

Converts a vector from little endian to the target’s endianness.

On little endian this is a no-op. On big endian the bytes are swapped.

pub fn from_be(x: Self) -> Self

Converts a vector from big endian to the target’s endianness.

On big endian this is a no-op. On little endian the bytes are swapped.

impl i8x64

pub fn count_ones(self) -> Self

Returns the number of ones in the binary representation of the lanes of self.

pub fn count_zeros(self) -> Self

Returns the number of zeros in the binary representation of the lanes of self.

pub fn leading_zeros(self) -> Self

Returns the number of leading zeros in the binary representation of the lanes of self.

pub fn trailing_zeros(self) -> Self

Returns the number of trailing zeros in the binary representation of the lanes of self.

impl i8x64

pub fn shuffle1_dyn<I>(self, indices: I) -> Self where
Self: Shuffle1Dyn<Indices = I>,

Shuffle vector elements according to indices.

impl i8x64

pub fn eq(self, other: Self) -> m8x64

Lane-wise equality comparison.

pub fn ne(self, other: Self) -> m8x64

Lane-wise inequality comparison.

pub fn lt(self, other: Self) -> m8x64

Lane-wise less-than comparison.

pub fn le(self, other: Self) -> m8x64

Lane-wise less-than-or-equals comparison.

pub fn gt(self, other: Self) -> m8x64

Lane-wise greater-than comparison.

pub fn ge(self, other: Self) -> m8x64

Lane-wise greater-than-or-equals comparison.

impl i8x64

pub fn partial_lex_ord(&self) -> LexicographicallyOrdered<i8x64>

Returns a wrapper that implements PartialOrd.

impl i8x64

pub fn lex_ord(&self) -> LexicographicallyOrdered<i8x64>

Returns a wrapper that implements Ord.

impl i8x64

pub fn bitmask(self) -> u64

Creates a bitmask with the MSB of each vector lane.

If the vector has less than 8 lanes, the bits that do not correspond to any vector lanes are cleared.