Type Definition packed_simd_2::m32x8

pub type m32x8 = Simd<[m32; 8]>;

A 256-bit vector mask with 8 m32 lanes.

Implementations

impl m32x8

pub const fn new( x0: bool, x1: bool, x2: bool, x3: bool, x4: bool, x5: bool, x6: bool, x7: bool ) -> Self

Creates a new instance with each vector elements initialized with the provided values.

pub const fn lanes() -> usize

Returns the number of vector lanes.

pub const fn splat(value: bool) -> Self

Constructs a new instance with each element initialized to value.

pub fn extract(self, index: usize) -> bool

Extracts the value at index.

Panics

If index >= Self::lanes().

pub unsafe fn extract_unchecked(self, index: usize) -> bool

Extracts the value at index.

Safety

If index >= Self::lanes() the behavior is undefined.

pub fn replace(self, index: usize, new_value: bool) -> Self

Returns a new vector where the value at index is replaced by new_value.

Panics

If index >= Self::lanes().

pub unsafe fn replace_unchecked(self, index: usize, new_value: bool) -> Self

Returns a new vector where the value at index is replaced by new_value.

Safety

If index >= Self::lanes() the behavior is undefined.

impl m32x8

pub fn and(self) -> bool

Lane-wise bitwise and of the vector elements.

Note: if the vector has one lane, the first element of the vector is returned.

pub fn or(self) -> bool

Lane-wise bitwise or of the vector elements.

Note: if the vector has one lane, the first element of the vector is returned.

pub fn xor(self) -> bool

Lane-wise bitwise xor of the vector elements.

Note: if the vector has one lane, the first element of the vector is returned.

impl m32x8

pub fn all(self) -> bool

Are all vector lanes true?

pub fn any(self) -> bool

Is any vector lane true?

pub fn none(self) -> bool

Are all vector lanes false?

impl m32x8

pub fn eq(self, other: Self) -> m32x8

Lane-wise equality comparison.

pub fn ne(self, other: Self) -> m32x8

Lane-wise inequality comparison.

pub fn lt(self, other: Self) -> m32x8

Lane-wise less-than comparison.

pub fn le(self, other: Self) -> m32x8

Lane-wise less-than-or-equals comparison.

pub fn gt(self, other: Self) -> m32x8

Lane-wise greater-than comparison.

pub fn ge(self, other: Self) -> m32x8

Lane-wise greater-than-or-equals comparison.

impl m32x8

pub fn select<T>(self, a: Simd<T>, b: Simd<T>) -> Simd<T>where T: SimdArray<NT = <[m32; 8] as SimdArray>::NT>,

Selects elements of a and b using mask.

The lanes of the result for which the mask is true contain the values of a. The remaining lanes contain the values of b.

impl m32x8

pub fn partial_lex_ord(&self) -> LexicographicallyOrdered<m32x8>

Returns a wrapper that implements PartialOrd.

impl m32x8

pub fn lex_ord(&self) -> LexicographicallyOrdered<m32x8>

Returns a wrapper that implements Ord.

impl m32x8

pub fn shuffle1_dyn<I>(self, indices: I) -> Selfwhere Self: Shuffle1Dyn<Indices = I>,

Shuffle vector elements according to indices.

impl m32x8

pub fn bitmask(self) -> u8

Creates a bitmask with the MSB of each vector lane.

If the vector has less than 8 lanes, the bits that do not correspond to any vector lanes are cleared.